+++ support/update.js
@@ -30,6 +30,9 @@ request(emojiSrc, function (err, response, body) {
     if (!def.emoji) { return; }
 
     def.aliases.forEach(function (alias) {
+      // Drop aliases without names (with names "uXXXX")
+      if (/^u[0-9a-b]{4,}$/i.test(alias)) return;
+
       emojies[alias] = def.emoji;
     });
   });
@@ -39,7 +42,7 @@ request(emojiSrc, function (err, response, body) {
   var visible = fs.readFileSync(p.join(__dirname, 'visible.txt'), 'utf8');
 
   var emoji_light = _.omitBy(emojies, function (val) {
-    return visible.indexOf(val) < 0;
+    return visible.indexOf(val.replace(/\uFE0F/g, '')) < 0;
   });
   fs.writeFileSync(p.join(__dirname, '../lib/data/light.json'), JSON.stringify(emoji_light, null, 2), 'utf8');