+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+1.4.0 / 2017-06-20
+------------------
+
+- Emoji list update (full set only).
+
+
+1.3.0 / 2016-10-04
+------------------
+
+- Emoji lists update.
+- Don't miss emojis with variation chars after.
+- Filter emojis with fake names ("uXXXX").
+
+
 1.2.0 / 2016-05-27
 ------------------